Bridge Scroll To Section
Summary:
Bridge scroll to section action on Screen List Views.

We do this by adding an `eventsController` on the `RecyclerView` underlying the `nt:screen:list`. This controller is stored on the `nt:screen:list`'s template.

In order to access the client controller on `nt:screen:list`, we had to augment this template which involved restructuring how templates were parsed in a screen.

Reviewed By: usikder

Differential Revision: D18143152

fbshipit-source-id: 24c7dc516ec680703f1ed1ec8ddbbf7b5cd7f006