fix debug hierarchy position bug
Summary:
This is dog science to me ¯\_(ツ)_/¯

It seems there are situations where multiple `DebugComponents` in hierarchy have the same `InternalNode`.
When that happens the position of the element may be incorrect as the bounds are read from `InternalNode` data.
By making the child position 0,0 by negating parent node we can mitigate this.

Note, it usually isn't a problem as position is 0,0 in most of these cases

Debuging WorkSpeed chat screen:
https://pxl.cl/PKC5
before: P120676060
after: P120676460

Reviewed By: marcelogomez

Differential Revision: D18255024

fbshipit-source-id: ee0817322b040fc668edad1fa94d3dcd73c82d42