Retry D18209497: Optimize visibility handlers dispatching for large component trees
Summary:
This revives D18209497 and fixes the problem we were seeing when it first landed.
The item was always marked for removal from the known visibility items on the MountState if it was previously fully visible, which triggered invibile event handlers on it.
On the next mount its visibility events are triggered again - in surfaces which triggered state updates in visibility handlers this was causing infinite loops because the visibility handlers kept getting called.

Reviewed By: astreet

Differential Revision: D18371579

fbshipit-source-id: 5f1c631e048c96a1749490c6982a20bfe598f128