Add support for logging Litho internal events into startup tracer
Summary:
Added the following events to be logged into startup tracer:

## init_range start/end

This is most straightforward, just add start/end that corresponds to `RecyclerBinder#initRange` (which is basically computing the layout of the first item synchronously in UI thread).

## first mount start/end

This one is a bit tricky that the event is happening within `LithoView`, however, `RecyclerBinder` can inform which one is the first mount in initial rendering. Using some state in `RecyclerBinder` and passing it to `LithoView` (in `onBindViewHolder` after which `Litho#mount` follows) makes it possible to detect first mount.

Reviewed By: colriot

Differential Revision: D18305345

fbshipit-source-id: 8db60bd178a143b76af8703a46b59ca4c3c06f91