Make sure circular RecyclerSpec's report an accurate item count for accessibility
Summary:
The story composers "Text Story" UI is one of the few places in the app that uses a circular Recycler view. These recycler views report a size of Integer.MAX_VALUE so that we can treat them as infinitely long. Unfortunately, that makes Talkback actually announce the list as containing 2147483674 items.

This fixes the issue by adding an accessibility delegate for any circular recycler view and setting an accurate column and row count.

{F221634702}

Reviewed By: astreet

Differential Revision: D18387886

fbshipit-source-id: af650a684344f8a60d2e80a1e044380a24ef564e