Set the background on Host Component to fix ripple effect issue. (#606)
Summary:
When the component needs a host view and it has a background, then instead of generating a layout output and a drawable component for the background drawable, set the background directly on the host view. For the components that doesn't need a host view, the background is mounted as any other drawable on the componentHost. This fixes the Ripple Effect not seen on Row or Column components.

<!-- Thanks for submitting a pull request! We appreciate you taking the time to work on these
changes. Please provide enough information so that others can review your pull request. The three
fields below are mandatory. -->

The issue which I am fixing here is same as reported in https://github.com/facebook/litho/issues/341.
Ripple Drawable gets clipped when set as a background for Row or Column components.
This is a follow up for https://github.com/facebook/litho/pull/601

## Changelog

Fix Ripple effects not to be clipped when set as background for Row or Column
Pull Request resolved: https://github.com/facebook/litho/pull/606

Test Plan: Ripple works now and it gets projected outside the Row or Column component.

Reviewed By: pasqualeanatriello

Differential Revision: D18456255

Pulled By: muraziz

fbshipit-source-id: 600f027a76afa870bad0679518f2c7533a681a78