Tag RenderInfo passed to ComponentWarmer
Summary:
When there's a race and ComponentWarmer#prepare gets scheduled after the RecyclerBinder has actually started computing layouts, then ComponentWarmer will end up creating a new ComponentTreeHolder which will calculate a layout.
If we tag the RenderInfo with the warmer tag, then the ComponentWarmer will get the same ComponentTreeHolder that is already computing layout and we avoid computing layouts unnecessarily.

Reviewed By: bnerbinavei

Differential Revision: D18479874

fbshipit-source-id: 2a2eed28311116271fbe505bf87774e84730e2f8