Use global keys instead of visibility output ids
Summary:
Previous the global keys being a thing we were computing long ids for VisibilityOutputs in a similar way to global keys - based on the position of the node in the tree.
Global keys represent the same thing and we can use them insted of generating separate ids for VisibilityOutputs.

Reviewed By: adityasharat

Differential Revision: D18530892

fbshipit-source-id: 8eb4a2edba9cb6b4d490d3f158e3e5b8e3449bbc