Use simple heuristic for MountSpec mountType inference in ABI mode
Summary:
In ABI mode, we don't have access to other modules, and therefore we can't just keep going up the hierarchy of types to determine whether we're mounting a drawable or a view.

Instead, we can just guess based on the name of the type that we do have. In general, people include the word `Drawable` in the name of any `Drawable` subclass, so that is the heuristic that we will use for now.

We then check when we do a non-ABI build to verify that our assumption was correct. If it wasn't, we can force the user to tell us the mount type, and we'll use that next time we do an ABI build.

Reviewed By: asp2insp

Differential Revision: D18597369

fbshipit-source-id: f8843cf900946857b12ce47f97de86dbd2eadac6