Fix Event and Trigger generation in ABI generation mode
Summary:
Previously we were just using a return type of VOID in all cases, but the return type is used in code generation and so we were ending up building against the wrong thing.

Instead, just assume that the method return type and the event return type match, and get the return type from the method (which we do have). If they don't match, that will be caught by validation when we do the full build of the target.

Reviewed By: asp2insp

Differential Revision: D18605887

fbshipit-source-id: 9a5a97412e0768bb541dcd8f4c2502b1fe459731