Make ComponentBodyGenerator compatible with RunMode.ABI
Summary: We can't call `typeSpec.isSubInterface` in ABI mode, which means that we can't calculate the level of a component in collections, when they are given as props or state or treeprops (e.g. `Prop Set<List<Component>>`). However, in ABI mode we don't actually need this information (since this is for generating code that the ABI doesn't need), so we can just skip the check and return `Comparable.OTHER` instead.

Reviewed By: asp2insp

Differential Revision: D18617460

fbshipit-source-id: 1c31f2786842223cebea3b21e32936a2e54cb075