Implements layout diffing with the InternalNode tree
Summary:
Implements layout diffing with the InternalNode tree. This diff essentially will make `LayoutState#mDiffNodeRoot == LayoutState#mLayoutRoot` and use the Internal Nodes for layout diffing. When enabled it will skip creating the DiffNode tree entirely.

In order for this to work, when the config is enabled the `InternalNode` tree will not be discarded after layout calculating. This will incur an increase in memory pressure.

Reviewed By: pasqualeanatriello

Differential Revision: D18427904

fbshipit-source-id: a829fcc896eaa76db052ad0917354137cc6ebe7e