Rename 'dp'/'sp'/'px' value fields to 'value'
Summary: It may be confusing to distinguish dp/sp value field from extension prop usage, which converts a number to a respective Dp/Sp value, when used on a variable name. I.e. `textSize.sp` - is it a value of `textSize: Sp` var or a conversion of a `textSize: Int` to an `Sp` value. Now I'll be easy to understand without IDE.

Reviewed By: astreet

Differential Revision: D18815693

fbshipit-source-id: 185ba01212d5c026c48d36cecca8772e62dc83c1