Use correct previous section
Summary:
When there are multiple updates triggered on a ComponentTree before onDataBound is called, the flipper sections plugin is not receiving the correct previous section, because it would become invalid between the first update and the time onDataBound is called.
This makes sure that the previous root passed to the plugin is the previous root on which the changesets can be applied to obtain the current root.

Reviewed By: pasqualeanatriello

Differential Revision: D18728078

fbshipit-source-id: e51c0378cac5c5e6f24f80e9f2ad73cbc6d106d0