"[litho][PR] Set the background & foreground on Host Component to fix ripple effect issue. (Retry of D18456255)"
Summary:
Original description:
When the component needs a host view and it has a background, then instead of generating a layout output and a drawable component for the background drawable, set the background directly on the host view. For the components that doesn't need a host view, the background is mounted as any other drawable on the componentHost. This fixes the Ripple Effect not seen on Row or Column components.

Previous attempt was rolled back (D18456255 / ) due to some issues in other apps that were relying on previous implementation.

That issue has been fixed (by unsetting padding in ComponentHost). Besides changes in D18456255 this diff also enable setting foreground directly on `ComponentHost`.

Reviewed By: pasqualeanatriello

Differential Revision: D19088301

fbshipit-source-id: 1c250e375a3c5a45828180ca29553be5b4bbc129