Better list-typing for DataDiffSectionComponent
Summary:
I noticed a few places in code where we cast lists like `ImmutableList<? extends SomeGraphQL>` to `ImmutableList` in order for it to work with `DataDiffSection.<SomeGraphQL>create(c)`

This shouldn't be necessary: you can't pass a wildcard list into an argument expecting a non-wildcard, but the reverse works just fine.

DataDiffSection never requires adding to this list or creating a new instance of the list, so typing the lists as bounded generics is safe (and I'd argue safer, as it means the ifnra never will be able to mutate the list).

Reviewed By: dschafer

Differential Revision: D19149641

fbshipit-source-id: 29ff67e051997d93e4fd44225ff5fd30a8b1b81c