Made varArg argument optional
Summary: Developers often omit optional flag from the varArg parameters, assuming it would be empty list if not provided. But in reality this leads to missing required prop crashes. Recent error analysis showed it could cause as much as 20% of top crashes. We are making varArg optional to align with Java varArg parameters behaviour.

Reviewed By: colriot

Differential Revision: D19376605

fbshipit-source-id: 3f7af659da77d3ce10c5a5733798b9abe0060874