Add LayoutCache to RenderCore
Summary:
A LayoutCache is accessible by a layout function to both read the LayoutResult coming from a given node in a previous layout computation and to write values for the next layout computation.
Values written will only be accessible to the next layout computation so we don't need to worry about thread safety and we don't have to copy things around too much.

Reviewed By: astreet

Differential Revision: D19517285

fbshipit-source-id: 9910dd4367690cf4cadda745fdd7bf944d3afa7c