IncrementalModule for incremental processing during mount
Summary:
A generic incremental processing util class. Has support for vertical processing and I will put out a separate diff for horizontal processing.
In the next diff we'll use it for incremental visibility processing.
The plan is to make it generic enough that we can move incremental mount on top of it and move the logic for incremental mount and visibility processing outside of MountState.

Reviewed By: pasqualeanatriello

Differential Revision: D19194907

fbshipit-source-id: a619a67b5ef75e6469a273ad470177a03fd3dbe3