Remove HostRenderUnit
Summary:
This is a pretty big change (and simplification) of the RenderCore API.
In substance we want to remove hostRenderUnit and only have one RenderUnit per node.
We can achieve the same wrapping behaviours by providing a wrapping RenderUnit instead (will do this in a followup).
This simplifies a bunch of things like:
- Bounds computations
- Extensions in Bloks
- We can remove the Node from the LayoutResult

Reviewed By: astreet

Differential Revision: D19601159

fbshipit-source-id: 199c8d111ed817ef2903b03181bbeef928255dfa