Prevent IncompatibleClassChangeError: ComponentUtils.hasEquivalentFieldsSummary
Summary:
PROBLEM
-------
This constitutes 89.2% of overall ComponentUtils.hasEquivalentFields crashes in production. This is a known error which exists in Android Lollypop which was fixed in later OS versions.

Details of the issue which was reported to Google :
https://issuetracker.google.com/issues/37045084
https://github.com/google/gson/issues/726

To summarise,  libraries which uses annotations is facing this intermittently. Google closed this saying it is infeasible to fix this in older OS versions.

Reviewed By: astreet

Differential Revision: D19835819

fbshipit-source-id: cebef218bd97531d8892e31e8d40654edaaa97de