Remove white space between content and shadows
Summary:
On some devices, we were seeing white space at the bottom and right sides of CardSpecs that have white clipping colors and non-white backgrounds in the card's content. The cause for this is that the shadow and clipping calculations are sometimes based off of different shadow sizes.

- CardSpec takes in an `elevation` prop, and calculates the shadow sizes to pass into CardClipSpec directly from the shadow calculation static methods (such as `getShadowBottom()`).
- However, a different shadow size is being saved in CardShadowDrawable. CardSpec passes its `elevation` to CardShadowSpec, which passes it to CardShadowDrawable. But CardShadowDrawable rounds the shadow size to the nearest even integer before saving it (https://fburl.com/diffusion/455ktwek).

This means that if the `elevation` value is not even, the shadow sizes used by the clipping and shadow components will be different. This diff changes the static method used by card clipping to be the same as the shadow calculation.

Concrete example:
- On Pixel: CardSpec's elevation is 11px, CardShadowDrawable's is 10px. The white space appears.
- On Pixel 2 XL: CardSpec's elevation is 14px, CardShadowDrawable's is 14px. The white space does not appear.

-----

I suspect this bug wasn't noticed/prioritized before because most CardSpecs have a white content background, so the white space at the bottom is not differentiable.

-----

Here is how the CardSpec previously looked on the Pixel device. You can see white space at the bottom and sides of the card:

| Solid non-white content background | Image as content background | White content background (no issues) |
| https://pxl.cl/10kZq | https://pxl.cl/10kZt | https://pxl.cl/10kZv |

-----

See T58728684 for more details.

Reviewed By: astreet

Differential Revision: D19867346

fbshipit-source-id: e22c3c35895d5a0605d9bf25a676e017502614be