Add test for StickHeader update on its Component's update
Summary:
In T61734597 it was uncovered that Sticky Headers update was supported only on Scroll/Viewport change. But when respective component was updated in-place while being visible, not layout update was performed. This test will actually test the case.
But for it to actually fail we also needed to fix fake implementation of `ComponentTreeHolder` which didn't invalidate tree as the real implementation does. These custom fakes are BAD.

Reviewed By: muraziz

Differential Revision: D19925280

fbshipit-source-id: ba1786e369bf3c6739aa6aadfa0ddb2e5c2c0a71