Fix Sections not updating layout for sticky items with indices outside range ratio
Summary:
In T61734597 an issue with sticky headers was uncovered. When we update a bunch of items in `RecyclerCollectionComponent` we actually calculate layout only for `[firstVisibleIndex - rangeRatio * numberOfItemsInViewport, lastVisibleIndex + rangeRatio * numberOfItemsInViewport]`. When we scroll far enough that Sticky component's index falls out of this range it stops updating.

This diff forcefully computes layout for all Sticky components that where already shown in the screen and computed before.

Reviewed By: muraziz

Differential Revision: D19883099

fbshipit-source-id: 635920b96722d156d44e5d1292068c2cdaa35e40