Fix view hierarchy dump for litho
Summary:
Fix identifying `LithoView` derived classes to execute litho specific dump logic.

The bug was that we only looked at classes inside the litho namespace, but custom views that extend LithoView were ignored not handel at the right level.

This is not perfect as we loose some of the hierarchy data. A deeper look needs to happen to understand how we can properly dump Views that are both litho and native.

For now at least we will have all the UI elements in the dump.

Differential Revision: D19902916

fbshipit-source-id: dd8d8a5ab50fb4d0a808a3f26e16d893015061f9