Introduce setVersionedRoot
Summary: We need root versioning when the componentTree is used inside another tree. In this case we can't simply rely on timing for which root was the last to be set as layout running on multiple threads could race on setting root on the nested tree. Having an explicit versioning for setRoot means that we can use that instead of "the last one to be set wins"

Reviewed By: mihaelao

Differential Revision: D20003622

fbshipit-source-id: 7c2914535a8b81c84379520d9e18724223beb27c