Fix view-hierarchy for litho with scroll
Summary:
If litho view is the first view in a native scrollable container we don't calculate its location correctly because we only take the left/top and ignore scrolling.

To handle all scenarios I calculate the location on the screen for the view and adjust it by the parent view, this way all offset are correctly handled whatever it's scroll or any other.

Reviewed By: scotthovestadt

Differential Revision: D20190544

fbshipit-source-id: 297d3b7a1d37167bc77e1d3646238adfc21f004e