Modifies the is equivalent to check for the root component in reconciliation.
Summary:
Modifies the is equivalent to check for the root component in reconciliation.

The previous implementation (D18783273) had to be backed out because it broke `Component#shouldUpdate()` for mount specs. Previously it used to call `Component#isEquivalentTo` which is overridden for mount specs but got replaced with the reflection based field comparisons. So the `isEquivalentTo` of mount specs was never called.

The original diff (D18783273) was intended to fix a bug where the `LayoutState#isReconcilable` wrongly returned `false` for root components which had modified state values. This diff fixes the bug again by checking equivalence without state containers only for this use case.

Reviewed By: mihaelao

Differential Revision: D20223355

fbshipit-source-id: 188ebce9422f8fa6f44384b4e4de8e02db9ce57f