Adds shadow offset support to CardShadowDrawable (#628)
Summary:
This change adds x and y shadow offset support to CardShadowDrawable and
CardShadowSpec. Clients can use these offsets to cast different types of
shadows to their cards. At the moment the offsets are hardcoded in
CardShadowDrawable: the x offset is zero and the y offset is half the
shadow size. This change removes this hard-coding from
CardShadowDrawable and exposes two new properties in CardShadowSpec
(shadowDx and shadowDy) to enable clients to set these offsets.

## Changelog

Added support for x and y shadow offsets to CardShadow.
Pull Request resolved: https://github.com/facebook/litho/pull/628

Test Plan:
I have manually tested this change on a simple UI that comprises a list of cards with each card containing an image on the left and a text on the right. I was able to change the x and y offsets for the card shadow and see the effects on the UI. Using this new functionality I was able to create shadows that come from different light sources above, below, at left and at right of the card.

Please see testing details below and let me know if any further testing is required to accept this PR.

#### shadowSize = 8 / shadowDx = 0 / shadowDy = 0
![800](https://user-images.githubusercontent.com/28489751/74651186-2f9f0780-517b-11ea-8435-be3135f1bb4b.png)

#### shadowSize = 8 / shadowDx = 2 / shadowDy = 0
![820](https://user-images.githubusercontent.com/28489751/74651275-53624d80-517b-11ea-976c-719a0b34294a.png)

#### shadowSize = 8 / shadowDx = 0 / shadowDy = 2
![802](https://user-images.githubusercontent.com/28489751/74651273-53624d80-517b-11ea-8c36-8dace867d309.png)

#### shadowSize = 8 / shadowDx = 2 / shadowDy = 2
![822](https://user-images.githubusercontent.com/28489751/74651276-53fae400-517b-11ea-8a2a-d61e593b0649.png)

#### shadowSize = 8 / shadowDx = -2 / shadowDy = 0
![8-20](https://user-images.githubusercontent.com/28489751/74651269-52312080-517b-11ea-8fab-fba98aab092e.png)

#### shadowSize = 8 / shadowDx = 0 / shadowDy = -2
![80-2](https://user-images.githubusercontent.com/28489751/74651270-52c9b700-517b-11ea-99c5-4297e8c0cf5b.png)

#### shadowSize = 8 / shadowDx = -2 / shadowDy = -2
![8-2-2](https://user-images.githubusercontent.com/28489751/74651263-52312080-517b-11ea-9fc1-0cbc1d40bc8e.png)

#### shadowSize = 8 / shadowDx = 2 / shadowDy = -2
![82-2](https://user-images.githubusercontent.com/28489751/74651272-52c9b700-517b-11ea-80d5-21140d7ef2f6.png)

Reviewed By: adityasharat, pasqualeanatriello

Differential Revision: D20303138

Pulled By: muraziz

fbshipit-source-id: cabf6d921c85c84638b24f0a8bb816b5e8a7b621