Made ComponentGenerateUtils#updateLayoutComponent method return void
Summary:
We want to generate a component if it does not exist yet during #update method call. Generation could take significant time (longer than few seconds). We don't want to block the main Events thread during that process, that's why we are making this method to return nothing (to make it async in the future).
User-facing notifications should still be triggered, but from another class.

Reviewed By: muraziz

Differential Revision: D20397629

fbshipit-source-id: cc85984238249cbd7059b0852bb76ae678a269f7