Change processVisibilityOutputs signature
Summary:
To test the incremental mount extension in Litho's mount state, we want to be able to call processVisibilityOutputs after mount - however mount resets the previousLocalVisibleRect and isDirty values when it's done.
This allows us to cache those values before mount is finished and call `processVisibilityOutputs` with the correct values for the current mount state.

Reviewed By: adityasharat

Differential Revision: D20535778

fbshipit-source-id: 06b5e04ea67f86bc0f023305f79098e5c948f1a1