Use LayoutState that matches latest StateHandler when performing reconciliation
Summary: There was a bug where sometimes we'd use the main-thread LayoutState for reconciliation which wasn't as up-to-date as the StateHandler we were passing in. The end result was the ability to accidentally drop state updates. The fix is to always use the LayoutState which was committed with the StateHandler that we will use.

Reviewed By: pasqualeanatriello

Differential Revision: D20620906

fbshipit-source-id: 4a5db828aab15ff63ff43ba2fe25ed14895a81f0