Extract mounted child LithoViews from MountDelegateTarget
Summary:
Part of preparation to delegate mount to RenderCore.MountState.
LithoView gets the list of mounted child LithoViews to do operations like setting visibility hint on all children LithoViews.
This adds an API to get the list of all mounted contents from a MountDelegateTarget - LithoView will use that to extract the child LithoViews.

Reviewed By: adityasharat

Differential Revision: D20624327

fbshipit-source-id: 4b2c4b413c3cf04d2f8ed3c642bb37761dc11a72