Sync calling requestRemeasure in RecyclerBinder from multiple threads
Summary:
requestRemeasure is always called on the UI thread except when it's called as part of a callback invoked from ComponentTree#setRootAndSizeSpecAsync. https://fburl.com/diffusion/k2am1ka9
In that case it can be called off the UI thread, and it's accessing the RecyclerView outside of a sync block while something could be unmouting the RecyclerBinder on a different thread: https://fburl.com/diffusion/1f74zlk5
This makes sure we sync around the requestRemeasure call when it can be accessed on a bg thread.

Reviewed By: arpitratan

Differential Revision: D20670332

fbshipit-source-id: 8565c45bc2330f945d4442b784095217d7286e93