Keep track of dirty mount in LithoView
Summary:
When using RenderCore.MountState we can't get the isDirty info from it - this replaces all mMountState.isDirty() calls with isMountStateDirty() and adds a new dirty flag on the LithoView directly. When we use RenderCore.MountState, we'll return mIsMountStateDirty from mIsMountStateDirty().
This does not change current behaviour - the flag will be used in the next diffs

Reviewed By: adityasharat

Differential Revision: D20646736

fbshipit-source-id: 8f515c34891f84e699f868349ba5f9a7496f21fb