Add mount(RenderTree) on MountDelegateTarget
Summary:
Adds `mount(RenderTree renderTree)` to MountDelegateTarget, to be called from the host (LithoView).
If we want to use Litho.MountState's mount with a RenderTree, it requires a lot of changes - adding the custom object on the RenderTree is a shortcut to get the LayoutState and use that to mount with Litho.MountState.
Alternatively I can also cast in LithoView to the appropriate MountState to call mount on so I don't *need* to add this if people feel strongly against it.

Reviewed By: adityasharat

Differential Revision: D20648099

fbshipit-source-id: 6b9230865c9cc1e5f552b8cd3ca25c058d044fde