Trigger callbacks on VisibilityOutputsExtension correctly from MountState
Summary:
In MountState it's possible to trigger a LithoView mount with disabled visibility processing https://fburl.com/diffusion/zmfy8rb2 when it's animating, and then mount it with enabling visibility processing.
If we check whether we should process visibility outputs or not when when calling beforeMount on the VisibilityOutputsExtension we won't initialise the input correctly in this case and we'll NPE when we actually want to process vis outputs.
This is only an issue when using the visibility extension directly from MountState; the version which we want to test eventually will always call these methods correctly from LithoView. This fixes the NPE in T64771905. Good news is that this wasn't a problem in the extension itself

Reviewed By: adityasharat

Differential Revision: D20771982

fbshipit-source-id: 37e54db1d325de3f926693b400c14a46556d9a50