Adds API to get LithoMountData from MountItem
Summary:
This diff adds an API to get LithoMountData from MountItem. This API mimic what the RenderCore's MountItem has.

If mount data is null while running Litho on RenderCore's MountState then the RuntimeException fail fast and implies that the MountItem was created by Litho, which should not happen when Litho is delegating mount to RenderCore.

Reviewed By: mihaelao

Differential Revision: D20816076

fbshipit-source-id: 820350426ff09f13427e3be3d9634c29bc2fe08e