NullPointerException fix for mComponentTree.getRecyclingMode
Summary:
This is a temporary fix till all the callers have moved to calling setComponentTree on main thread. LithoView.setComponentTree is not designed to be thread safe.
Main problem is that mComponentTree is set to null from a different thread and it crashes when the current thread tries to call any method on it.

Using function local variable instead of mComponentTree to reduce the number of times mComponentTree is accessed. This would not have happened at the first place if setComponentTree was called only from UI thread. getRecyclingMode is experimental and would be removed in future hence this code will automatically be removed in future.

Reviewed By: muraziz

Differential Revision: D20918930

fbshipit-source-id: 1ba120192848d1a4b0ac9a2e536cb9eee8fce8fc