Makes RenderCore's MountItem and some APIs public.
Summary:
Makes some APIs of RenderCore's MountItem public and the class itself public; so that, when this replaces Litho's MountItem there are no compile errors.

This diff also adds a new method `setHost` which is required for the Litho RenderCore interop. Litho reset the `Host` on a `MountItem` in `MountState#remountComponentHostToRootIfNeeded` for disappearing items.

Reviewed By: mihaelao

Differential Revision: D20837570

fbshipit-source-id: 861bcf4620027e9be1c9c257037038d1626744b9