Creates HostComponent with a ComponentContext
Summary:
HostComponent is created without a ComponentContext, but that is required because `LithoRenderUnit` Binders needs the `ComponentContext` to call `mount` of the `HostComponent` when Litho will delegate mounting to RenderCore's `MountState`.

From `LithoRenderUnit.LithoMountBinder`

```
public static class LithoMountBinder implements Binder<LithoRenderUnit, Object> {
  Override
    public void bind(Context context, Object content, LithoRenderUnit unit, Object data) {
      LayoutOutput output = unit.output;

      // >>> IN ORDER TO CALL Component#mount() the ComponentContext is required <<<
      output.getComponent().mount(output.getComponent().getScopedContext(), content);
    }
}
```

This diff now creates HostComponent with a ComponentContext.

Reviewed By: mihaelao

Differential Revision: D20597205

fbshipit-source-id: 082f9c2cb886c2b17ddf028872a2c3567ee08422