Make sure to always set text property on AccessibilityNodeInfo for TextSpec virtual views
Summary:
TextSpec's that have AccessibleClickableSpan spannables as part of their text could cause errors if the contentDescription property of the AccessibleClickableSpan was set to null.

When you set the `accessibleClickableSpans` prop of TextSpec to true, Litho creates virtual children for accessibility, but these children require some text to be set on them.  Since AccessibleClickableSpan can technically have a null contentDescription, it is possible that no text was being set at all, and when this happens the `ExploreByTouchHelper` class that is backing the `ComponentAccessibilityDelegate` throws this error:

https://github.com/elastic/AndroidSDKMirror-28/blob/master/androidx/customview/widget/ExploreByTouchHelper.java#L803

The fix here is easy. We should *always* set the text property to the value of the spanned text, and only set the contentDescription property if one exists. The contentDescription will be announced by Talkback if it exists, but if it doesn't it will fall back to reading the contents of the spanned text itself.

Differential Revision: D20903132

fbshipit-source-id: 1d987f01568d5e7e8ed419b58daa4b533019bc0a