Reorder the unbind bind calls in MountState#updateMountItemIfNeeded
Summary:
Prior to this diff, the order of unbinding and rebinding the `Binders` of a `RenderUnit` when a `MountItem` is updated was as follows:

```
1. unbind ATTACH binders

2. rebind ATTACH binders

3. unbind MOUNT binders

4. rebind MOUNT binders
```

This is not semantically sound; ideally the sequence should be:

```
1. unbind ATTACH binders

2. unbind MOUNT binders

3. rebind MOUNT binders

4. rebind ATTACH binders
```

Primarily because when a `MountItem` in mounted and then unmounted the order is:

```
OnMount():
  1. bind MOUNT binders
  2. bind ATTACH binders

OnUnmount()
  1. unbind ATTACH binders
  2. unbind MOUNT binders
```

The "remount" should basically be the reverse of this order; which would be the same as the ideal order mentioned above. This diff changes `MountState#updateMountItemIfNeeded` to achieve the ideal order.

In addition, to be more semantically correct  `Binder#unbind` on the binders of a `RenderUnit` should be called in the opposite order from `Binder#bind`.

Reviewed By: astreet

Differential Revision: D20645401

fbshipit-source-id: 9eb2d0ff30e1b62891f6800a81da4002e3fc2097