Clear ComponentTree in IncrementalMountHelper when it's being released.
Summary:
# Problem

When `IncrementalMountHelper#release()` is being called, the `ViewPager` listener is only removed from the `ViewPager` at the end UI thread run-loop.

This means that even if the current backing `LithoView` is already detached, `ViewPagerListener#onPageScrolled()` will still be called for the current run-loop, triggering an incremental mount on the existing `ComponentTree`.

The side effect of this is visibility events such as `onVisible` and `onFocus` are being fired when the UI is not longer visible

# Solution
We remove the reference to the `ComponentTree` when the listener is being released. This operation should be cheap enough to be run synchronously on the UI thread.

Reviewed By: mihaelao

Differential Revision: D20953593

fbshipit-source-id: cb96c39d9bff004249ef11e6d1c6e189fa914d12