Call suppressInvalidationsOnHosts from onViewOffset
Summary:
When using incremental mount extension, when mounting hosts through onViewOffset they are created with `mSuppressInvalidations` enabled, but after they are mounted it is not set to false. This happens in control because `suppressInvalidationsOnHosts` is called on every mount.
With the extension, `suppressInvalidationsOnHosts` was only called at the end of a dirty mount, which means that hosts which were created during incremental mount were never getting invalidated and drawn.

This calls `suppressInvalidationsOnHosts` when mount is triggered by an extension (this would happen on an incremental mount step). It's different from what control does because it calls it for every mounted host rather than at the end of the incremental mount step, but it's a no-op if nothing changes for a host.
The alternative is to add another callback on the MountDelegate to notify it that incremental mount is done and call this, but I think in practice this doesn't add much overhead compared to what was happening before - let me know your thoughts.

Reviewed By: pasqualeanatriello

Differential Revision: D21226363

fbshipit-source-id: e583c8cd383457b2376f49e15f0094bda240f9e0